+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.8.0 (Unpublished)
 * Add `OrdinalNumberPrinter` to print ordinal numbers: 1st, 2nd, 3rd, 4th, ...
 * Add `Printer.switcher` to use different printers based on conditions.
+* Add `Printer.where` to condition the printing on the printed object.
 * Add `Printer` support for `take`, `skip`, `takeLast` and `skipLast`.
 * Add `Printer` support for `truncateRight`, `truncateLeft`, and `truncateCenter` to also operate on word and sentence boundaries. Include the ellipsis in the width constraint.
 * Add `String` extensions `take`, `takeTo`, `skip`, `skipTo`, `takeLast`, `takeLastTo`, `skipLast`, and `skipLastTo`.